backend/ml_models/shuttle_run/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.23.2
numpy==1.26.0
pandas==2.2.3
psycopg2-binary==2.9.9
python-dotenv==1.0.0
opencv-python==4.10.0.84
tensorflow==2.18.0
joblib==1.4.2
sqlalchemy==2.0.21
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.9
